Separated the Schedule component from ScheduleContainer.
